# services/subscription-external/docs/swagger.yaml
basePath: /
definitions:
...
  monitoring.AlertThresholds:
    properties:
      database_errors:
        description: Alert if > 10 in 1 hour
        type: integer
      high_failure_rate:
        description: Alert if > 80%
        type: number
      high_queue_size:
        description: Alert if > 10000
        type: integer
      low_success_rate:
        description: Alert if < 60%
        type: number
      processing_delay:
        description: Alert if > 5 minutes
        type: number
      service_unavailable:
        description: Alert if service down
        type: boolean
    type: object
...
host: localhost:8083
info:
  contact:
    email: support@omni-connect.com
    name: API Support
    url: https://omni-connect.com/support
  description: '|'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://omni-connect.com/terms/
  title: TIMWE Subscription External Service API
  version: "1.0"
paths:
...
  /api/v1/subscription-external/monitoring/alerts:
    get:
      consumes:
      - application/json
      description: Get current monitoring alerts with optional filtering by severity,
        type, and acknowledgment status
      parameters:
      - description: Filter by alert severity (low, medium, high, critical)
        in: query
        name: severity
        type: string
      - description: Filter by alert type
        in: query
        name: type
        type: string
      - description: Filter by acknowledgment status
        in: query
        name: acknowledged
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Alerts retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get monitoring alerts
      tags:
      - Monitoring
  /api/v1/subscription-external/monitoring/alerts/acknowledge:
    post:
      consumes:
      - application/json
      description: Mark a monitoring alert as acknowledged by the specified user
      parameters:
      - description: ID of the alert to acknowledge
        in: query
        name: alert_id
        required: true
        type: string
      - description: Identifier of who acknowledged the alert
        in: query
        name: acknowledged_by
        type: string
      - description: Optional notes about the acknowledgment
        in: query
        name: notes
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Alert acknowledged successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - missing alert_id
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Acknowledge monitoring alert
      tags:
      - Monitoring
  /api/v1/subscription-external/monitoring/alerts/clear:
    post:
      consumes:
      - application/json
      description: Remove old acknowledged alerts older than the specified hours
      parameters:
      - description: 'Hours threshold for clearing alerts (default: 24)'
        in: query
        name: hours
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Alerts cleared successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Clear monitoring alerts
      tags:
      - Monitoring
  /api/v1/subscription-external/monitoring/dashboard:
    get:
      consumes:
      - application/json
      description: Get comprehensive monitoring dashboard data including metrics,
        alerts, and charts
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard data retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get monitoring dashboard data
      tags:
      - Monitoring
...
  /api/v1/subscription-external/monitoring/thresholds:
    put:
      consumes:
      - application/json
      description: Update the alert thresholds for charging failures
      parameters:
      - description: New alert thresholds
        in: body
        name: thresholds
        required: true
        schema:
          $ref: '#/definitions/monitoring.AlertThresholds'
      produces:
      - application/json
      responses:
        "200":
          description: Thresholds updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid threshold data
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update monitoring alert thresholds
      tags:
      - Monitoring
...
schemes:
- http
- https
swagger: "2.0"